Added main to create dataset_file and compute_image_normalizations

+++ pagesegmentation/scripts/compute_image_normalizations.py
@@ -7,27 +7,6 @@ import multiprocessing
 import tqdm
 import json
 
-parser = argparse.ArgumentParser()
-parser.add_argument("--input_dir", type=str, required=True,
-                    help="Image directory to process")
-parser.add_argument("--average_all", action="store_true", default=True,
-                    help="bla")
-parser.add_argument("--cut_left", type=float, default=0.05)
-parser.add_argument("--cut_right", type=float, default=0.05)
-parser.add_argument("--inverse", action="store_false", default=True)
-parser.add_argument("--output_dir", type=str, required=True,
-                    help="The output dir for the info files")
-
-args = parser.parse_args()
-
-if not os.path.exists(args.input_dir):
-    raise Exception("Cannot open {}".format(args.input_dir))
-
-files = [os.path.join(args.input_dir, f) for f in os.listdir(args.input_dir)]
-
-# files = files[:10]
-
-
 def computeCharHeight(file_name):
     if not os.path.exists(file_name):
         raise Exception("File does not exist at {}".format(file_name))
@@ -57,27 +36,50 @@ def computeCharHeight(file_name):
     except:
         return None
 
+def main():
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--input_dir", type=str, required=True,
+                        help="Image directory to process")
+    parser.add_argument("--average_all", action="store_true", default=True,
+                        help="bla")
+    parser.add_argument("--cut_left", type=float, default=0.05)
+    parser.add_argument("--cut_right", type=float, default=0.05)
+    parser.add_argument("--inverse", action="store_false", default=True)
+    parser.add_argument("--output_dir", type=str, required=True,
+                        help="The output dir for the info files")
 
-with multiprocessing.Pool(processes=12) as p:
-    char_heights = [v for v in tqdm.tqdm(p.imap(computeCharHeight, files), total=len(files))]
+    args = parser.parse_args()
 
-if args.average_all:
-    av_height = np.mean([c for c in char_heights if c])
-    char_heights = [av_height] * len(char_heights)
+    if not os.path.exists(args.input_dir):
+        raise Exception("Cannot open {}".format(args.input_dir))
 
-if args.output_dir and not os.path.exists(args.output_dir):
-    os.mkdir(args.output_dir)
+    files = [os.path.join(args.input_dir, f) for f in os.listdir(args.input_dir)]
 
-for file, height in zip(files, char_heights):
-    filename, file_extension = os.path.splitext(os.path.basename(file))
-    if args.output_dir:
-        output_file = os.path.join(args.output_dir, filename + ".norm")
-        with open(output_file, 'w') as f:
-            json.dump(
-                {"file": file, "char_height": int(height)},
-                f,
-                indent=4
-            )
+    # files = files[:10]
 
-    print(file, height)
 
+    with multiprocessing.Pool(processes=12) as p:
+        char_heights = [v for v in tqdm.tqdm(p.imap(computeCharHeight, files), total=len(files))]
+
+    if args.average_all:
+        av_height = np.mean([c for c in char_heights if c])
+        char_heights = [av_height] * len(char_heights)
+
+    if args.output_dir and not os.path.exists(args.output_dir):
+        os.mkdir(args.output_dir)
+
+    for file, height in zip(files, char_heights):
+        filename, file_extension = os.path.splitext(os.path.basename(file))
+        if args.output_dir:
+            output_file = os.path.join(args.output_dir, filename + ".norm")
+            with open(output_file, 'w') as f:
+                json.dump(
+                    {"file": file, "char_height": int(height)},
+                    f,
+                    indent=4
+                )
+
+        print(file, height)
+
+if __name__ == '__main__':
+    main()
